Move pages and window to separate folder, rename to app
