--- development/tk_helloW.py
@@ -1,8 +1,0 @@
-import tkinter as tk
-
-root = tk.Tk()
-
-w = tk.Label(root, text="Hello Tkinter!")
-w.pack()
-
-root.mainloop()
