--- development/tk_helloW_OOP.py
@@ -1,38 +1,0 @@
-"""from tkinter import *
-from tkinter import ttk
-
-root = ttk.Tk()
-
-label = ttk.Label(root, text="Hello world!")
-
-label.pack()
-
-root.mainloop()"""
-
-import tkinter as tk
-
-
-class Application(tk.Frame):
-    def __init__(self, master=None):
-        super().__init__(master)
-        self.pack()
-        self.create_widgets()
-
-    def create_widgets(self):
-        self.hi_there = tk.Button(self)
-        self.hi_there["text"] = "Hello World\n(click me)"
-        self.hi_there["command"] = self.say_hi
-        self.hi_there.pack(side="top")
-
-        self.quit = tk.Button(
-            self, text="QUIT", fg="red", command=root.destroy
-        )
-        self.quit.pack(side="bottom")
-
-    def say_hi(self):
-        print("hi there, everyone!")
-
-
-root = tk.Tk()
-app = Application(master=root)
-app.mainloop()
