+++ pyproject.toml
@@ -7,7 +7,7 @@ include = '\.pyi?$'
 line_length = 79
 multi_line_output = 3
 include_trailing_comma = true
-known_third_party = ["backend", "database", "entities", "matplotlib", "numpy"]
+known_third_party = ["database", "entities", "matplotlib", "numpy", "pages"]
 
 [tool.flake8]
 max-line-length = 79
